fix: Switch to AudioCenter implemented in bones - remove from Jumpy

+++ src/settings.rs
@@ -26,11 +26,11 @@ fn load_settings(game: &mut Game) {
 pub struct Settings {
     /// The main scaling factor for all game audios. This is done on top of the
     /// scaling factor specific to the audio type.
-    pub main_volume: f64,
+    pub main_volume: f32,
     /// The scaling factor for music.
-    pub music_volume: f64,
+    pub music_volume: f32,
     /// The scaling factor for sound effects.
-    pub effects_volume: f64,
+    pub effects_volume: f32,
     /// Whether to display the game fullscreen.
     pub fullscreen: bool,
     /// The player controller bindings
